Remove Generator class and add SourceDataset class
