+++ src/milkstraw_client/__init__.py
@@ -1,7 +1,7 @@
 import os
 
 from milkstraw_client.api_client import APIClient
-from milkstraw_client.generator import Generator
+from milkstraw_client.source_dataset import SourceDataset
 
 edge_service_url = os.environ.get("MILKSTRAW_EDGE_SERVICE_URL")
 if edge_service_url is None:
